requirements.txt
azure-functions==1.7.2
pandas
requests
azure.cosmos